test(tui): consolidate viewport goldens
